evaluation/alternativeClusteringApproaches/preAnalysisByAnLLM/requirements.txt
joblib
backoff
numpy
pandas
umap-learn
hdbscan
streamlit
plotly
tqdm
python-dotenv
google.generativeai
google.api_core